+++ src/test_type/fortune.rs
@@ -14,7 +14,7 @@ use std::cmp::min;
 const FORTUNES: &str = "<!doctype html><html><head><title>Fortunes</title></head><body><table><tr><th>id</th><th>message</th></tr><tr><td>11</td><td>&lt;script&gt;alert(&quot;This should not be displayed in a browser alert box.&quot;);&lt;/script&gt;</td></tr><tr><td>4</td><td>A bad random number generator: 1, 1, 1, 1, 1, 4.33e+67, 1, 1, 1</td></tr><tr><td>5</td><td>A computer program does what you tell it to do, not what you want it to do.</td></tr><tr><td>2</td><td>A computer scientist is someone who fixes things that aren&apos;t broken.</td></tr><tr><td>8</td><td>A list is only as strong as its weakest link. — Donald Knuth</td></tr><tr><td>0</td><td>Additional fortune added at request time.</td></tr><tr><td>3</td><td>After enough decimal places, nobody gives a damn.</td></tr><tr><td>7</td><td>Any program that runs right is obsolete.</td></tr><tr><td>10</td><td>Computers make very fast, very accurate mistakes.</td></tr><tr><td>6</td><td>Emacs is a nice operating system, but I prefer UNIX. — Tom Christaensen</td></tr><tr><td>9</td><td>Feature: A bug with seniority.</td></tr><tr><td>1</td><td>fortune: No such file or directory</td></tr><tr><td>12</td><td>フレームワークのベンチマーク</td></tr></table></body></html>";
 
 pub struct Fortune {
-    pub concurrency_levels: Vec<usize>,
+    pub concurrency_levels: Vec<u32>,
     pub database_verifier: Box<dyn DatabaseInterface>,
 }
 impl Executor for Fortune {
@@ -85,7 +85,7 @@ impl Executor for Fortune {
     }
 }
 impl Fortune {
-    fn get_wrk_command(&self, url: &str, duration: usize, concurrency: usize) -> Vec<String> {
+    fn get_wrk_command(&self, url: &str, duration: u32, concurrency: u32) -> Vec<String> {
         vec![
             "wrk",
             "-H",
@@ -102,7 +102,7 @@ impl Fortune {
             "--timeout",
             "8",
             "-t",
-            &format!("{}", min(concurrency, num_cpus::get())),
+            &format!("{}", min(concurrency, num_cpus::get() as u32)),
             url,
         ].iter().map(|item| item.to_string()).collect()
     }
