+++ src/test_type/query/cached_query.rs
@@ -8,7 +8,7 @@ use crate::verification::Messages;
 use std::cmp::min;
 
 pub struct CachedQuery {
-    pub concurrency_levels: Vec<usize>,
+    pub concurrency_levels: Vec<u32>,
     pub database_verifier: Box<dyn DatabaseInterface>,
 }
 impl Query for CachedQuery {}
@@ -59,7 +59,7 @@ impl Executor for CachedQuery {
     }
 }
 impl CachedQuery {
-    fn get_wrk_command(&self, url: &str, duration: usize, concurrency: usize) -> Vec<String> {
+    fn get_wrk_command(&self, url: &str, duration: u32, concurrency: u32) -> Vec<String> {
         vec![
             "wrk",
             "-H",
@@ -76,7 +76,7 @@ impl CachedQuery {
             "--timeout",
             "8",
             "-t",
-            &format!("{}", min(concurrency, num_cpus::get())),
+            &format!("{}", min(concurrency, num_cpus::get() as u32)),
             url,
         ].iter().map(|item| item.to_string()).collect()
     }
